// frontend/UNSERfrontend/src/app/api/model/createTodoDto.ts
/**
 * Campus Manager
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateTodoDto { 
    name: string;
    fertig: boolean;
    abgabedatum: string;
}